refactor(weight-chart): move WeightChart component to route folder
